refactor: move SelectAreaLayer to map/layers folder
